adds status to the new router
